RM-2346: 7_7 to master merge

+++ sugarcrm/modules/pmse_Business_Rules/clients/base/views/dashlet-businessrules/dashlet-businessrules.js
@@ -65,7 +65,28 @@
 //    },
     businessRulesLayout: function (model) {
         var redirect = model.module+"/"+model.id+"/layout/businessrules";
-        app.router.navigate(redirect , {trigger: true, replace: true });
+        var verifyURL = app.api.buildURL(
+                'pmse_Project',
+                'verify',
+                {id: model.get('id')},
+                {baseModule: this.module}),
+            self = this;
+        app.api.call('read', verifyURL, null, {
+            success: function(data) {
+                if (!data) {
+                    app.router.navigate(redirect , {trigger: true, replace: true });
+                } else {
+                    app.alert.show('business-rule-design-confirmation',  {
+                        level: 'confirmation',
+                        messages: App.lang.get('LBL_PMSE_PROCESS_BUSINESS_RULES_EDIT', model.module),
+                        onConfirm: function () {
+                            app.router.navigate(redirect , {trigger: true, replace: true });
+                        },
+                        onCancel: $.noop
+                    });
+                }
+            }
+        });
     },
 
     /**
@@ -146,19 +167,38 @@
      * @param {String} params.module Module name.
      */
     deleteRecord: function(model) {
-        var self = this;
-        this._modelToDelete = true;
-        var name = model.get('name') || '',
-            context = app.lang.get('LBL_MODULE_NAME_SINGULAR', model.module).toLowerCase() + ' - ' + name.trim();
-        app.alert.show(model.get('id') + ':deleted', {
-            level: 'confirmation',
-            messages: app.utils.formatString(app.lang.get('LBL_PRO_DELETE_CONFIRMATION', model.module)),
-            onConfirm: function() {
-                model.destroy({
-                    showAlerts: true,
-//                    relate: true
-                    success: self._getRemoveRecord()
-                });
+        var verifyURL = app.api.buildURL(
+                'pmse_Project',
+                'verify',
+                {id: model.get('id')},
+                {baseModule: this.module}),
+            self = this;
+        this._modelToDelete = model;
+        app.api.call('read', verifyURL, null, {
+            success: function(data) {
+                if (!data) {
+                    app.alert.show('delete_confirmation', {
+                        level: 'confirmation',
+                        messages: app.utils.formatString(app.lang.get('LBL_PRO_DELETE_CONFIRMATION', model.module)),
+                        onConfirm: function () {
+                            model.destroy({
+                                showAlerts: true,
+                                success: self._getRemoveRecord()
+                            });
+                        },
+                        onCancel: function () {
+                            self._modelToDelete = null;
+                        }
+                    });
+                } else {
+                    app.alert.show('message-id', {
+                        level: 'warning',
+                        title: app.lang.get('LBL_WARNING'),
+                        messages: app.lang.get('LBL_PMSE_PROCESS_BUSINESS_RULES_DELETE', model.module),
+                        autoClose: false
+                    });
+                    self._modelToDelete = null;
+                }
             }
         });
     },
